These had an incomplete implementation before because the one from cppreference.com didn't work in C++11.
Now with C++14, we can make the full implementation from cppreference.com work and have c++17-compatible verisons of std::apply/std::invoke.

Differential Revision: [D18869642](https://our.internmc.facebook.com/intern/diff/D18869642/)

[ghstack-poisoned]

+++ caffe2/contrib/fakelowp/test/test_layernorm_nnpi_fp16.py
@@ -101,8 +101,8 @@ class LayerNorm(serial.SerializedTestCase):
         workspace.RunNet(pred_net_onnxified.name)
         Y_glow = workspace.FetchBlob("Y")
 
-        if not np.allclose(Y_glow.astype(np.float16), Y_c2.astype(np.float16)):
-            diff_Y = np.abs(Y_glow - Y_c2).astype(np.float16)
+        if not np.allclose(Y_glow, Y_c2):
+            diff_Y = np.abs(Y_glow - Y_c2)
             print_test_debug_info(
                 "layernorm",
                 {
